requirements.txt
hydralit_components==1.0.10
mysql-connector-python==8.1.0
numpy==1.24.3
pandas==1.5.3
Requests==2.31.0
streamlit==1.28.2
streamlit-aggrid==0.3.4.post3
streamlit_folium==0.15.1
streamlit_option_menu==0.3.6
plotly==5.18.0
mysqlclient==2.2.1
